Initial checkin of MicroPython code
The MicroPython code for the 8266 board. So far it's a WiFi web server, able to serve files from a designated part of the filesystem. It presents a directory page showing all the files, when a query doesn't match an existing file. It responds with a simplified directory listing when a magic URL is accessed, intended to be parsed by other badges.