Add serial receive machinery.
Serial frames start and end with a STX character 0x02 (^B).
The last four bytes before the final STX are a hex
CRC-CCITT (XMODEM) checksum of the other bytes.

CRC16 code copied under GPL3 from
https://github.com/gtrafimenkov/pycrc16

Import usocket instead of socket, to be explicit.

Initialize the UART before the web server, so the web server's
signon message doesn't get stomped by the UART re-init.

Testing shows that reasonable size UART transmissions complete
in one go, never short-sending. That may not be true if there
is handshaking involved, but we will assume it for now.